Use redis to store map of hash to url
In order to address the problem caused by lengthy urls, this stores a
sha1 hash in redis as a key with the url as the value. This hash is used
in place of the base64 of the url. A new route has been added at
`/url/<hash>` for cantaloupe to turn the hash back into a url.

Note that this change will not work until cataloupe has also been
changed to get the url from the hash.